stats and various minor
