+++ examples/ui/public/index.html
@@ -6,6 +6,10 @@
     margin: 0;
     padding: 0;
   }
+  main {
+    font-family: sans-serif;
+    padding: 2em;
+  }
 </style>
 <div id='menu'></div>
 <main></main>
