+++ examples/ui/src/api/index.js
@@ -1,95 +1,18 @@
-import queryString from 'query-string'
+import GrpcClient from './grpcClient'
+import Stack from './stack'
+import Topic from './topic'
 
-function sleep (ms) {
-  return new Promise((resolve, reject) => {
-    setTimeout(resolve, ms)
-  })
-}
-
-class Stack {
-  constructor (data, api) {
-    this.id = data.id
-    this.name = data.name
-    this.state = data.state
-    this.api = api
-  }
-  async tasks () {
-    const data = await this.api.getJson(`stack/${this.id}/tasks`)
-    return data.message
-  }
-  async logs () {
-    const data = await this.api.logs({
-      stack: this.name
-    })
-    return data
-  }
-  async start () {
-    const results = await this.api.postJson(`stack/${this.id}/start`)
-    await sleep(100)
-    this.state = 'Running'
-    return results
-  }
-  async stop () {
-    const results = this.api.postJson(`stack/${this.id}/stop`)
-    await sleep(100)
-    this.state = 'Stopped'
-    return results
-  }
-  async remove () {
-    const results = await this.api.deleteJson(`stack/${this.id}`)
-    await sleep(100)
-    return results
-  }
-}
-
-class Topic {
-  constructor (data, api) {
-    this.id = data.id
-    this.name = data.name
-    this.api = api
-  }
-  async remove () {
-    const results = await this.api.deleteJson(`topic/${this.id}`)
-    await sleep(100)
-    return results
-  }
-}
-
-export default class AmpApi {
+export default class AmpApi extends GrpcClient {
   constructor (base) {
+    super()
     this.base = base || 'http://amplifier-api.local.appcelerator.io/v1/'
-  }
-  async getJson (path, query) {
-    if (query) {
-      path += '?' + queryString.stringify(query)
-    }
-    const request = await fetch(this.base + path)
-    const json = await request.json()
-    return json
-  }
-  async postJson (path, body, query) {
-    if (query) {
-      path += '?' + queryString.stringify(query)
-    }
-    const request = await fetch(this.base + path, {
-      method: 'POST',
-      body: body ? JSON.stringify(body) : undefined
-    })
-    const json = await request.json()
-    return json
-  }
-  async deleteJson (path, query) {
-    if (query) {
-      path += '?' + queryString.stringify(query)
-    }
-    const request = await fetch(this.base + path, {
-      method: 'DELETE'
-    })
-    const json = await request.json()
-    return json
   }
   async logs (query) {
     const data = await this.getJson('log', query)
+    return data.entries || []
+  }
+  async stats (query) {
+    const data = await this.getJson('stats', query)
     return data.entries || []
   }
   async stacks (query) {
