+++ examples/ui/src/api/sleep.js
@@ -1,0 +1,5 @@
+export default function sleep (ms) {
+  return new Promise((resolve, reject) => {
+      setTimeout(resolve, ms)
+  })
+}
