+++ examples/ui/src/components/CodeMirrorYaml.html
@@ -1,0 +1,32 @@
+<link rel="stylesheet" href="/codemirror.css">
+<div ref:code></div>
+
+<script>
+  import CodeMirror from 'codemirror'
+  import 'codemirror/mode/yaml/yaml.js'
+
+  export default {
+    onrender () {
+      this.cm = CodeMirror(this.refs.code, {
+        value: this.get('yaml'),
+        mode: 'yaml'
+      })
+      this.cm.on('changes', () => {
+        const yaml = this.cm.doc.getValue()
+        this.set({yaml})
+      })
+    },
+
+    data () {
+      return {
+        yaml: '',
+      };
+    },
+
+    methods: {
+      getCodeMirror () {
+        return this.cm
+      }
+    }
+  };
+</script>
