+++ examples/ui/src/components/DisplayError.html
@@ -1,0 +1,17 @@
+<div>{{message}}</div>
+
+<script>
+  export default {
+    data () {
+      return {
+        message: ''
+      }
+    }
+  }
+</script>
+
+<style>
+  div {
+    color: indianred;
+  }
+</style>
