+++ examples/ui/src/components/StackEdit.html
@@ -1,20 +1,17 @@
-<link rel="stylesheet" href="/codemirror.css">
-<input bind:value='name'>
-<div id='code'></div>
-{{#if stackSpec.parseError}}
-  <div class='error'>
-    {{stackSpec.parseError.message}}
-    {{stackSpec.parseError.parsedLine ? `on line ${stackSpec.parseError.parsedLine}` : ''}}
-  </div>
+<h1>Stack Edit</h1>
+<label>Name: <input bind:value='name'></label>
+<CodeMirrorYaml bind:yaml></CodeMirrorYaml>
+{{#if error}}
+  <DisplayError message='{{error}}'></DisplayError>
 {{else}}
   <button on:click='create()'>Create</button>
 {{/if}}
 
 
 <script>
-  import CodeMirror from 'codemirror'
-  import 'codemirror/mode/yaml/yaml.js'
-  import yaml from 'yamljs'
+  import CodeMirrorYaml from './CodeMirrorYaml.html'
+  import DisplayError from './DisplayError.html'
+  import yamlParser from 'yamljs'
   import AmpApi from '../api/index.js'
   const api = new AmpApi()
 
@@ -35,27 +32,22 @@
     image: redis`
 
   export default {
-    onrender () {
-      this.cm = CodeMirror(document.querySelector('#code'), {
-        value: this.get('stackyaml'),
-        mode: 'yaml'
-      })
-      this.cm.on('changes', () => {
-        this.set({stackyaml: this.cm.doc.getValue()})
-      })
+    components: {
+      CodeMirrorYaml,
+      DisplayError,
     },
 
     data () {
       return {
-        name: 'myCounter',
-        stackyaml: stackyaml
+        name: 'counter',
+        yaml: stackyaml,
       };
     },
 
     computed: {
-      stackSpec: (name, stackyaml) => {
+      stackSpec: (name, yaml) => {
         try {
-          const spec = yaml.parse(stackyaml)
+          const spec = yamlParser.parse(yaml)
           const stackSpec = {
             name,
             services: [],
@@ -82,12 +74,14 @@
             }
             stackSpec.services.push(serviceSpec)
           }
-          console.log(stackSpec)
           return stackSpec
         } catch (parseError) {
-          console.log(parseError)
           return {parseError}
         }
+      },
+      error (stackSpec) {
+        return stackSpec.parseError ?
+          `${stackSpec.parseError.message} on line ${stackSpec.parseError.parsedLine}` : ''
       }
     },
 
@@ -100,10 +94,14 @@
 </script>
 
 <style>
-  .error {
-    color: indianred;
-  }
   button {
     cursor: pointer;
+  }
+  label, input {
+    vertical-align: middle;
+  }
+  input {
+    padding: 0.25em;
+    border-radius: 0.5em;
   }
 </style>
