+++ examples/ui/src/components/Stacks.html
@@ -88,13 +88,13 @@
 </script>
 
 <style>
-  body {
-    padding: 2em;
-  }
   table {
     text-align: left;
     font-family: monospace;
     border-collapse: collapse;
+  }
+  td {
+    padding-right: 1em;
   }
   tbody tr {
     cursor: pointer;
